Phone : Fix for setting audio mode during call
CRs fixed : 191643

When the call is initiated using dialer both speaker and mic are unmuted. But
when the call is made through call manager in QXDM, STK call setup and AT
command, the audio mode is not set. Hence no voice is available.This fix sets
the audio mode when there is call state change and phone is in OFFHOOK state.